Move top level functions to Node
