+++ pkl-core/src/main/java/org/pkl/core/stdlib/syntax/SyntaxNodes.java
@@ -15,22 +15,16 @@
  */
 package org.pkl.core.stdlib.syntax;
 
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.dsl.Specialization;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import org.jspecify.annotations.Nullable;
-import org.pkl.core.ast.lambda.ApplyVmFunction1Node;
 import org.pkl.core.runtime.Identifier;
 import org.pkl.core.runtime.SyntaxModule;
-import org.pkl.core.runtime.VmFunction;
 import org.pkl.core.runtime.VmList;
 import org.pkl.core.runtime.VmNull;
-import org.pkl.core.runtime.VmPair;
 import org.pkl.core.runtime.VmTyped;
 import org.pkl.core.runtime.VmUtils;
-import org.pkl.core.stdlib.ExternalMethod2Node;
 import org.pkl.core.stdlib.VmObjectFactory;
 import org.pkl.parser.syntax.generic.FullSpan;
 import org.pkl.parser.syntax.generic.Node;
@@ -71,58 +65,8 @@ public final class SyntaxNodes {
                       : VmNull.withoutDefault())
           .addTypedProperty("span", nd -> nd.spanVm);
 
-  public abstract static class walk extends ExternalMethod2Node {
-    @Child private ApplyVmFunction1Node applyVisit = ApplyVmFunction1Node.create();
-
-    @Specialization
-    @TruffleBoundary
-    protected VmTyped eval(VmTyped self, VmTyped node, VmFunction visit) {
-      var result = walkNode(node, visit);
-      // the root of the returned tree has no parent
-      if (result.hasExtraStorage()) {
-        ((NodeData) result.getExtraStorage()).parentVm = null;
-      }
-      return result;
-    }
-
-    private VmTyped walkNode(VmTyped nodeVm, VmFunction visit) {
-      var visited = applyVisit.execute(visit, nodeVm);
-
-      VmTyped node;
-      boolean descend;
-      if (visited instanceof VmPair pair) {
-        node = (VmTyped) pair.getFirst();
-        descend = (Boolean) pair.getSecond();
-      } else {
-        // `null`: leave this node unchanged and keep descending
-        node = nodeVm;
-        descend = true;
-      }
-      if (!descend) {
-        return node;
-      }
-
-      var childrenVm = (VmList) VmUtils.readMember(node, Identifier.CHILDREN);
-      var length = childrenVm.getLength();
-      if (length == 0) {
-        return node;
-      }
-
-      var newChildren = new Object[length];
-      var changed = false;
-      for (var i = 0; i < length; i++) {
-        var child = (VmTyped) childrenVm.get(i);
-        var newChild = walkNode(child, visit);
-        newChildren[i] = newChild;
-        changed |= newChild != child;
-      }
-      // reuse the node (and its extra storage) untouched when nothing below changed
-      return changed ? rebuild(node, newChildren) : node;
-    }
-  }
-
   /** Rebuild a node from {@code template} (its type, span, text) with new children. */
-  private static VmTyped rebuild(VmTyped template, Object[] newChildrenVm) {
+  static VmTyped rebuild(VmTyped template, Object[] newChildrenVm) {
     var nodeType =
         NodeType.valueOf(
             ((String) VmUtils.readMember(template, Identifier.TYPE)).toUpperCase(Locale.ROOT));
