+++ pkl-core/src/test/files/LanguageSnippetTests/input/syntax/traversal.pkl
@@ -23,15 +23,15 @@ local sample: syntax.Node =
 
 facts {
   ["fold counts nodes by predicate"] {
-    syntax.fold(sample, 0, (acc, n) -> if (n.type == "import") acc + 1 else acc) == 2
+    sample.fold(0, (acc, n) -> if (n.type == "import") acc + 1 else acc) == 2
 
-    syntax.fold(sample, 0, (acc, n) -> if (n.type == "if_expr") acc + 1 else acc) == 1
+    sample.fold(0, (acc, n) -> if (n.type == "if_expr") acc + 1 else acc) == 1
 
-    syntax.fold(sample, 0, (acc, n) -> if (n.type == "when_generator") acc + 1 else acc) == 0
+    sample.fold(0, (acc, n) -> if (n.type == "when_generator") acc + 1 else acc) == 0
   }
 
   ["fold accumulates into a collection"] {
-    syntax.fold(sample, List(), (acc, n) -> if (n.type == "identifier") acc.add(n.text) else acc)
+    sample.fold(List(), (acc, n) -> if (n.type == "identifier") acc.add(n.text) else acc)
       == List(
         "baz",
         "Point",
@@ -52,14 +52,14 @@ facts {
 
   ["fold visits a node before its children (pre-order)"] {
     // the module node is visited first
-    syntax.fold(sample, List(), (acc, n) -> acc.add(n.type)).first == "module"
+    sample.fold(List(), (acc, n) -> acc.add(n.type)).first == "module"
   }
 
   ["descendants enumerates the whole tree"] {
-    syntax.descendants(sample).first == sample
+    sample.descendants().first == sample
 
-    syntax.descendants(sample) == syntax.fold(sample, List(), (acc, n) -> acc.add(n))
+    sample.descendants() == sample.fold(List(), (acc, n) -> acc.add(n))
 
-    syntax.descendants(sample).filter((n) -> n.type == "class").length == 1
+    sample.descendants().filter((n) -> n.type == "class").length == 1
   }
 }
